Выделил package screens для экранов приложения
Также выделил package bottomNavigation
