+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/bottomNavigation/Tab.kt
@@ -1,0 +1,11 @@
+package com.designdrivendevelopment.kotelok.screens.bottomNavigation
+
+import androidx.fragment.app.Fragment
+
+data class Tab(
+    val name: String,
+    private val hostFragmentInstantiation: () -> Fragment
+) {
+    val hostFragment: Fragment by lazy { hostFragmentInstantiation.invoke() }
+    val tag = "OPEN_" + name.uppercase()
+}
